requirements.txt
django
django-debug-toolbar
falco-app